// lib/company/generated/researchSteps.ts
export const RESEARCH_TARGETS: Array<{ name: string; priority: Priority }> = [
  {
    "name": "TCS",
    "priority": 1
  },
  {
    "name": "Infosys",
    "priority": 1
  },
  {
    "name": "Wipro",
    "priority": 1
  },
  {
    "name": "Cognizant",
    "priority": 1
  },
  {
    "name": "Accenture",
    "priority": 1
  },
  {
    "name": "Capgemini",
    "priority": 1
  },
  {
    "name": "HCLTech",
    "priority": 1
  },
  {
    "name": "Tech Mahindra",
    "priority": 1
  },
  {
    "name": "LTIMindtree",
    "priority": 1
  },
  {
    "name": "Deloitte",
    "priority": 1
  },
  {
    "name": "IBM",
    "priority": 1
  },
  {
    "name": "DXC Technology",
    "priority": 1
  },
  {
    "name": "Mphasis",
    "priority": 1
  },
  {
    "name": "Hexaware",
    "priority": 1
  },
  {
    "name": "Persistent Systems",
    "priority": 1
  },
  {
    "name": "Google",
    "priority": 2
  },
  {
    "name": "Microsoft",
    "priority": 2
  },
  {
    "name": "Amazon",
    "priority": 2
  },
  {
    "name": "Meta",
    "priority": 2
  },
  {
    "name": "Apple",
    "priority": 2
  },
  {
    "name": "Adobe",
    "priority": 2
  },
  {
    "name": "LinkedIn",
    "priority": 2
  },
  {
    "name": "NVIDIA",
    "priority": 2
  },
  {
    "name": "Uber",
    "priority": 2
  },
  {
    "name": "Atlassian",
    "priority": 2
  },
  {
    "name": "Salesforce",
    "priority": 2
  },
  {
    "name": "Oracle",
    "priority": 2
  },
  {
    "name": "Flipkart",
    "priority": 2
  },
  {
    "name": "PhonePe",
    "priority": 2
  },
  {
    "name": "Razorpay",
    "priority": 2
  },
  {
    "name": "Swiggy",
    "priority": 3
  },
  {
    "name": "Zomato",
    "priority": 3
  },
  {
    "name": "Meesho",
    "priority": 3
  },
  {
    "name": "CRED",
    "priority": 3
  },
  {
    "name": "Groww",
    "priority": 3
  },
  {
    "name": "Zepto",
    "priority": 3
  },
  {
    "name": "Dream11",
    "priority": 3
  },
  {
    "name": "Paytm",
    "priority": 3
  },
  {
    "name": "Juspay",
    "priority": 3
  },
  {
    "name": "Ola",
    "priority": 3
  },
  {
    "name": "Zoho",
    "priority": 3
  },
  {
    "name": "Freshworks",
    "priority": 3
  },
  {
    "name": "Postman",
    "priority": 3
  },
  {
    "name": "BrowserStack",
    "priority": 3
  },
  {
    "name": "Chargebee",
    "priority": 3
  },
  {
    "name": "Druva",
    "priority": 3
  },
  {
    "name": "Myntra",
    "priority": 3
  },
  {
    "name": "InMobi",
    "priority": 3
  },
  {
    "name": "Navi",
    "priority": 3
  },
  {
    "name": "Udaan",
    "priority": 3
  }
]